Hungarian Planner
Hungarian Planner uses the Hungarian Algorithm (padding with centroids
or zeros depending on ambulances availability) to dispatch the most
suitable ambulances to their destinations.

Needs commenting and testing

Also we need to handle the previous plans (do we keep them partially eg
only ambulances with a patient or drop them completely? where do the
patients already on ambulances go in this case?)